add manual data pipeline from alphavantage api to mysql
